+++ jib-core/src/main/java/com/google/cloud/tools/jib/blob/FileBlob.java
@@ -16,7 +16,7 @@
 
 package com.google.cloud.tools.jib.blob;
 
-import com.google.cloud.tools.jib.hash.DigestUtil;
+import com.google.cloud.tools.jib.hash.Digests;
 import java.io.BufferedInputStream;
 import java.io.IOException;
 import java.io.InputStream;
@@ -36,7 +36,7 @@ class FileBlob implements Blob {
   @Override
   public BlobDescriptor writeTo(OutputStream outputStream) throws IOException {
     try (InputStream fileIn = new BufferedInputStream(Files.newInputStream(file))) {
-      return DigestUtil.computeDigest(fileIn, outputStream);
+      return Digests.computeDigest(fileIn, outputStream);
     }
   }
 }