+++ jib-core/src/main/java/com/google/cloud/tools/jib/builder/steps/BuildResult.java
@@ -17,10 +17,9 @@
 package com.google.cloud.tools.jib.builder.steps;
 
 import com.google.cloud.tools.jib.blob.BlobDescriptor;
-import com.google.cloud.tools.jib.hash.DigestUtil;
+import com.google.cloud.tools.jib.hash.Digests;
 import com.google.cloud.tools.jib.image.DescriptorDigest;
 import com.google.cloud.tools.jib.image.Image;
-import com.google.cloud.tools.jib.image.Layer;
 import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
 import com.google.cloud.tools.jib.image.json.ImageToJsonTranslator;
 import com.google.common.io.ByteStreams;
@@ -38,8 +37,7 @@ public class BuildResult {
    * @return a new {@link BuildResult} with the image's digest and id
    * @throws IOException if writing the digest or container configuration fails
    */
-  static BuildResult fromImage(
-      Image<Layer> image, Class<? extends BuildableManifestTemplate> targetFormat)
+  static BuildResult fromImage(Image image, Class<? extends BuildableManifestTemplate> targetFormat)
       throws IOException {
     ImageToJsonTranslator imageToJsonTranslator = new ImageToJsonTranslator(image);
     BlobDescriptor containerConfigurationBlobDescriptor =
@@ -49,7 +47,7 @@ public class BuildResult {
     BuildableManifestTemplate manifestTemplate =
         imageToJsonTranslator.getManifestTemplate(
             targetFormat, containerConfigurationBlobDescriptor);
-    DescriptorDigest imageDigest = DigestUtil.computeJsonDigest(manifestTemplate);
+    DescriptorDigest imageDigest = Digests.computeJsonDigest(manifestTemplate);
     DescriptorDigest imageId = containerConfigurationBlobDescriptor.getDigest();
     return new BuildResult(imageDigest, imageId);
   }