+++ jib-core/src/main/java/com/google/cloud/tools/jib/builder/steps/PullBaseImageStep.java
@@ -23,6 +23,7 @@ import com.google.cloud.tools.jib.builder.BuildStepType;
 import com.google.cloud.tools.jib.builder.ProgressEventDispatcher;
 import com.google.cloud.tools.jib.builder.TimerEventDispatcher;
 import com.google.cloud.tools.jib.builder.steps.PullBaseImageStep.BaseImageWithAuthorization;
+import com.google.cloud.tools.jib.cache.CacheCorruptedException;
 import com.google.cloud.tools.jib.configuration.BuildConfiguration;
 import com.google.cloud.tools.jib.configuration.ImageConfiguration;
 import com.google.cloud.tools.jib.configuration.credentials.Credential;
@@ -31,16 +32,17 @@ import com.google.cloud.tools.jib.event.events.ProgressEvent;
 import com.google.cloud.tools.jib.http.Authorization;
 import com.google.cloud.tools.jib.image.DescriptorDigest;
 import com.google.cloud.tools.jib.image.Image;
-import com.google.cloud.tools.jib.image.Layer;
+import com.google.cloud.tools.jib.image.ImageReference;
 import com.google.cloud.tools.jib.image.LayerCountMismatchException;
 import com.google.cloud.tools.jib.image.LayerPropertyNotFoundException;
 import com.google.cloud.tools.jib.image.json.BadContainerConfigurationFormatException;
+import com.google.cloud.tools.jib.image.json.BuildableManifestTemplate;
 import com.google.cloud.tools.jib.image.json.ContainerConfigurationTemplate;
 import com.google.cloud.tools.jib.image.json.JsonToImageTranslator;
+import com.google.cloud.tools.jib.image.json.ManifestAndConfig;
 import com.google.cloud.tools.jib.image.json.ManifestTemplate;
 import com.google.cloud.tools.jib.image.json.UnknownManifestFormatException;
 import com.google.cloud.tools.jib.image.json.V21ManifestTemplate;
-import com.google.cloud.tools.jib.image.json.V22ManifestTemplate;
 import com.google.cloud.tools.jib.json.JsonTemplateMapper;
 import com.google.cloud.tools.jib.registry.RegistryAuthenticator;
 import com.google.cloud.tools.jib.registry.RegistryClient;
@@ -51,6 +53,7 @@ import com.google.common.util.concurrent.ListenableFuture;
 import com.google.common.util.concurrent.ListeningExecutorService;
 import com.google.common.util.concurrent.MoreExecutors;
 import java.io.IOException;
+import java.util.Optional;
 import java.util.concurrent.Callable;
 import java.util.concurrent.ExecutionException;
 import javax.annotation.Nullable;
@@ -64,17 +67,16 @@ class PullBaseImageStep
   /** Structure for the result returned by this step. */
   static class BaseImageWithAuthorization {
 
-    private final Image<Layer> baseImage;
+    private final Image baseImage;
     private final @Nullable Authorization baseImageAuthorization;
 
     @VisibleForTesting
-    BaseImageWithAuthorization(
-        Image<Layer> baseImage, @Nullable Authorization baseImageAuthorization) {
+    BaseImageWithAuthorization(Image baseImage, @Nullable Authorization baseImageAuthorization) {
       this.baseImage = baseImage;
       this.baseImageAuthorization = baseImageAuthorization;
     }
 
-    Image<Layer> getBaseImage() {
+    Image getBaseImage() {
       return baseImage;
     }
 
@@ -107,8 +109,8 @@ class PullBaseImageStep
   @Override
   public BaseImageWithAuthorization call()
       throws IOException, RegistryException, LayerPropertyNotFoundException,
-          LayerCountMismatchException, ExecutionException,
-          BadContainerConfigurationFormatException {
+          LayerCountMismatchException, ExecutionException, BadContainerConfigurationFormatException,
+          CacheCorruptedException {
     // Skip this step if this is a scratch image
     ImageConfiguration baseImageConfiguration = buildConfiguration.getBaseImageConfiguration();
     if (baseImageConfiguration.getImage().isScratch()) {
@@ -126,6 +128,10 @@ class PullBaseImageStep
                 "Getting base image "
                     + buildConfiguration.getBaseImageConfiguration().getImage()
                     + "..."));
+
+    if (buildConfiguration.isOffline()) {
+      return new BaseImageWithAuthorization(pullBaseImageOffline(), null);
+    }
 
     try (ProgressEventDispatcher progressEventDispatcher =
             progressEventDispatcherFactory.create(
@@ -210,7 +216,7 @@ class PullBaseImageStep
    * @throws BadContainerConfigurationFormatException if the container configuration is in a bad
    *     format
    */
-  private Image<Layer> pullBaseImage(
+  private Image pullBaseImage(
       @Nullable Authorization registryAuthorization,
       ProgressEventDispatcher progressEventDispatcher)
       throws IOException, RegistryException, LayerPropertyNotFoundException,
@@ -228,19 +234,24 @@ class PullBaseImageStep
     switch (manifestTemplate.getSchemaVersion()) {
       case 1:
         V21ManifestTemplate v21ManifestTemplate = (V21ManifestTemplate) manifestTemplate;
+        buildConfiguration
+            .getBaseImageLayersCache()
+            .writeMetadata(
+                buildConfiguration.getBaseImageConfiguration().getImage(), v21ManifestTemplate);
         return JsonToImageTranslator.toImage(v21ManifestTemplate);
 
       case 2:
-        V22ManifestTemplate v22ManifestTemplate = (V22ManifestTemplate) manifestTemplate;
-        if (v22ManifestTemplate.getContainerConfiguration() == null
-            || v22ManifestTemplate.getContainerConfiguration().getDigest() == null) {
+        BuildableManifestTemplate buildableManifestTemplate =
+            (BuildableManifestTemplate) manifestTemplate;
+        if (buildableManifestTemplate.getContainerConfiguration() == null
+            || buildableManifestTemplate.getContainerConfiguration().getDigest() == null) {
           throw new UnknownManifestFormatException(
-              "Invalid container configuration in Docker V2.2 manifest: \n"
-                  + JsonTemplateMapper.toUtf8String(v22ManifestTemplate));
+              "Invalid container configuration in Docker V2.2/OCI manifest: \n"
+                  + JsonTemplateMapper.toUtf8String(buildableManifestTemplate));
         }
 
         DescriptorDigest containerConfigurationDigest =
-            v22ManifestTemplate.getContainerConfiguration().getDigest();
+            buildableManifestTemplate.getContainerConfiguration().getDigest();
 
         try (ProgressEventDispatcherWrapper progressEventDispatcherWrapper =
             new ProgressEventDispatcherWrapper(
@@ -257,10 +268,49 @@ class PullBaseImageStep
           ContainerConfigurationTemplate containerConfigurationTemplate =
               JsonTemplateMapper.readJson(
                   containerConfigurationString, ContainerConfigurationTemplate.class);
-          return JsonToImageTranslator.toImage(v22ManifestTemplate, containerConfigurationTemplate);
+          buildConfiguration
+              .getBaseImageLayersCache()
+              .writeMetadata(
+                  buildConfiguration.getBaseImageConfiguration().getImage(),
+                  buildableManifestTemplate,
+                  containerConfigurationTemplate);
+          return JsonToImageTranslator.toImage(
+              buildableManifestTemplate, containerConfigurationTemplate);
         }
     }
 
     throw new IllegalStateException("Unknown manifest schema version");
+  }
+
+  /**
+   * Retrieves the cached base image.
+   *
+   * @return the cached image
+   * @throws IOException when an I/O exception occurs
+   * @throws CacheCorruptedException if the cache is corrupted
+   * @throws LayerPropertyNotFoundException if adding image layers fails
+   * @throws BadContainerConfigurationFormatException if the container configuration is in a bad
+   *     format
+   */
+  private Image pullBaseImageOffline()
+      throws IOException, CacheCorruptedException, BadContainerConfigurationFormatException,
+          LayerCountMismatchException {
+    ImageReference baseImage = buildConfiguration.getBaseImageConfiguration().getImage();
+    Optional<ManifestAndConfig> metadata =
+        buildConfiguration.getBaseImageLayersCache().retrieveMetadata(baseImage);
+    if (!metadata.isPresent()) {
+      throw new IOException(
+          "Cannot run Jib in offline mode; " + baseImage + " not found in local Jib cache");
+    }
+
+    ManifestTemplate manifestTemplate = metadata.get().getManifest();
+    if (manifestTemplate instanceof V21ManifestTemplate) {
+      return JsonToImageTranslator.toImage((V21ManifestTemplate) manifestTemplate);
+    }
+
+    ContainerConfigurationTemplate configurationTemplate =
+        metadata.get().getConfig().orElseThrow(IllegalStateException::new);
+    return JsonToImageTranslator.toImage(
+        (BuildableManifestTemplate) manifestTemplate, configurationTemplate);
   }
 }