+++ jib-gradle-plugin/src/main/java/com/google/cloud/tools/jib/gradle/JibExtension.java
@@ -17,6 +17,7 @@
 package com.google.cloud.tools.jib.gradle;
 
 import com.google.cloud.tools.jib.plugins.common.PropertyNames;
+import java.io.File;
 import org.gradle.api.Action;
 import org.gradle.api.Project;
 import org.gradle.api.model.ObjectFactory;
@@ -48,8 +49,8 @@ import org.gradle.api.tasks.Optional;
  *     format = OCI
  *     appRoot = '/app'
  *   }
- *   extraDirectory {
- *     path = file('path/to/extra/dir')
+ *   extraDirectories {
+ *     paths = ['/path/to/extra/dir', 'can/be/relative/to/project/root']
  *     permissions = [
  *       '/path/on/container/file1': 744,
  *       '/path/on/container/file2': 123
@@ -67,9 +68,11 @@ public class JibExtension {
   private final BaseImageParameters from;
   private final TargetImageParameters to;
   private final ContainerParameters container;
-  private final ExtraDirectoryParameters extraDirectory;
-
+  private final ExtraDirectoriesParameters extraDirectories;
   private final Property<Boolean> allowInsecureRegistries;
+
+  @Deprecated boolean extraDirectoryConfigured;
+  @Deprecated boolean extraDirectoriesConfigured;
 
   public JibExtension(Project project) {
     ObjectFactory objectFactory = project.getObjects();
@@ -77,7 +80,7 @@ public class JibExtension {
     from = objectFactory.newInstance(BaseImageParameters.class);
     to = objectFactory.newInstance(TargetImageParameters.class);
     container = objectFactory.newInstance(ContainerParameters.class);
-    extraDirectory = objectFactory.newInstance(ExtraDirectoryParameters.class, project);
+    extraDirectories = objectFactory.newInstance(ExtraDirectoriesParameters.class, project, this);
 
     allowInsecureRegistries = objectFactory.property(Boolean.class);
 
@@ -97,19 +100,22 @@ public class JibExtension {
     action.execute(container);
   }
 
-  public void extraDirectory(Action<? super ExtraDirectoryParameters> action) {
-    action.execute(extraDirectory);
+  @Deprecated
+  public void extraDirectory(Action<? super ExtraDirectoriesParameters> action) {
+    extraDirectoryConfigured = true;
+    action.execute(extraDirectories);
   }
 
-  /**
-   * Sets extra directory paths. {@code extraDirectories} can be any suitable object describing file
-   * paths convertible by {@link Project#files} (such as {@code List<File>}).
-   *
-   * @param extraDirectories paths to set.
-   */
-  // non-plural to retain backward-compatibility for the "jib.extraDirectory" config parameter
-  public void setExtraDirectory(Object extraDirectories) {
-    this.extraDirectory.setPath(extraDirectories);
+  public void extraDirectories(Action<? super ExtraDirectoriesParameters> action) {
+    extraDirectoriesConfigured = true;
+    action.execute(extraDirectories);
+  }
+
+  @Deprecated
+  // for the deprecated "jib.extraDirectory" config parameter
+  public void setExtraDirectory(File extraDirectory) {
+    extraDirectoryConfigured = true;
+    this.extraDirectories.setPath(extraDirectory);
   }
 
   public void setAllowInsecureRegistries(boolean allowInsecureRegistries) {
@@ -134,10 +140,17 @@ public class JibExtension {
     return container;
   }
 
+  @Deprecated
   @Nested
   @Optional
-  public ExtraDirectoryParameters getExtraDirectory() {
-    return extraDirectory;
+  public ExtraDirectoriesParameters getExtraDirectory() {
+    return extraDirectories;
+  }
+
+  @Nested
+  @Optional
+  public ExtraDirectoriesParameters getExtraDirectories() {
+    return extraDirectories;
   }
 
   @Input