+++ jib-plugins-common/src/main/java/com/google/cloud/tools/jib/plugins/common/ProjectProperties.java
@@ -72,4 +72,6 @@ public interface ProjectProperties {
   String getVersion();
 
   int getMajorJavaVersion();
+
+  boolean isOffline();
 }